[go-support] having the shopping cart in a separate package. travis build and TCK configuration in place
