+++ tck/src/it/resources/application.conf
@@ -1,32 +1,4 @@
-cloudstate-tck.combinations = [
-{
-  name = "Akka + Go"
-  tck {
-    hostname = "0.0.0.0"
-    port = 8090
-  }
-  proxy {
-    hostname = "127.0.0.1"
-    port = 9000
-    directory = ${user.dir}
-    command = ["java","-Xmx512M", "-Xms128M", "-Dconfig.resource=in-memory.conf", "-Dcloudstate.proxy.dev-mode-enabled=true", "-jar", "proxy/core/target/scala-2.12/akka-proxy.jar"]
-
-    env-vars {
-      USER_FUNCTION_PORT = "8090"
-    }
-  }
-  frontend {
-    hostname = "127.0.0.1"
-    port = 8080
-    directory = ${user.dir}/go-support
-    command = ["./gosupport"]
-    env-vars {
-      HOST = "127.0.0.1"
-      PORT = "8080"
-    }
-  }
-}
-,{
+cloudstate-tck.combinations = [{
   name = "Akka + Node.js"
   tck {
     hostname = "0.0.0.0"
@@ -61,8 +33,8 @@ cloudstate-tck.combinations = [
       DEBUG = "cloudstate-*"
     }
   }
-}
-,{
+},
+{
   name = "Akka + Java"
   tck {
     hostname = "0.0.0.0"
@@ -89,5 +61,31 @@ cloudstate-tck.combinations = [
       PORT = "8088"
     }
   }
-}
-]
+},
+{
+  name = "Akka + Go"
+  tck {
+    hostname = "0.0.0.0"
+    port = 8090
+  }
+  proxy {
+    hostname = "127.0.0.1"
+    port = 9000
+    directory = ${user.dir}
+    command = ["java","-Xmx512M", "-Xms128M", "-Dconfig.resource=in-memory.conf", "-Dcloudstate.proxy.dev-mode-enabled=true", "-jar", "proxy/core/target/scala-2.12/akka-proxy.jar"]
+
+    env-vars {
+      USER_FUNCTION_PORT = "8090"
+    }
+  }
+  frontend {
+    hostname = "127.0.0.1"
+    port = 8080
+    directory = ${user.dir}/go-support/shoppingcart
+    command = ["./shoppingcart"]
+    env-vars {
+      HOST = "127.0.0.1"
+      PORT = "8080"
+    }
+  }
+}]
